parser 파일 분할 및 norminette
