+++ Makefile
@@ -6,7 +6,7 @@
 #    By: sungjpar <sungjpar@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/03/10 00:54:23 by sungjpar          #+#    #+#              #
-#    Updated: 2022/08/17 20:20:27 by sungjpar         ###   ########.fr        #
+#    Updated: 2022/08/18 14:51:40 by sungjpar         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -44,6 +44,8 @@ SRC_FILES = $(addprefix $(SRC_DIR),\
 			parser/lexer.c\
 			parser/lexer_utils.c\
 			parser/parser.c\
+			parser/parser_utils.c\
+			parser/parser_tester.c\
 )
 
 OBJS = ${SRC_FILES:.c=.o}
